+++ packages/components/autocomplete/src/use-autocomplete.tsx
@@ -2,7 +2,7 @@ import type { CSSUIObject, HTMLUIProps, UIPropGetter } from "@yamada-ui/core"
 import { layoutStyleProperties } from "@yamada-ui/core"
 import type { FormControlOptions } from "@yamada-ui/form-control"
 import {
-  getFormControlProperties,
+  formControlProperties,
   useFormControlProps,
 } from "@yamada-ui/form-control"
 import type { MotionUIPropGetter } from "@yamada-ui/motion"
@@ -374,9 +374,9 @@ export const useAutocomplete = <T extends string | string[] = string>({
 
   const { id } = rest
 
-  const formControlProps = pickObject(
+  const { "aria-readonly": _ariaReadonly, ...formControlProps } = pickObject(
     rest,
-    getFormControlProperties({ omit: ["aria-readonly"] }),
+    formControlProperties,
   )
   const [containerProps, inputProps] = splitObject<Dict, string>(
     omitObject(rest, [...popoverProperties, "onKeyDown", "onFocus"]),
@@ -1215,6 +1215,7 @@ export const useAutocompleteInput = () => {
     isAllSelected,
     formControlProps,
     inputProps,
+    isOpen,
   } = useAutocompleteContext()
 
   useUpdateEffect(() => {
@@ -1225,6 +1226,13 @@ export const useAutocompleteInput = () => {
     (props = {}, ref = null) => ({
       ref: mergeRefs(inputRef, ref),
       ...formControlProps,
+      role: "combobox",
+      "aria-haspopup": "listbox",
+      "aria-autocomplete": "list",
+      "aria-expanded": isOpen,
+      autoCapitalize: "none",
+      autoComplete: "off",
+      spellCheck: "false",
       ...inputProps,
       ...props,
       id,
@@ -1245,6 +1253,7 @@ export const useAutocompleteInput = () => {
       ),
     }),
     [
+      isOpen,
       inputProps,
       inputRef,
       formControlProps,