+++ packages/components/breadcrumb/src/breadcrumb.tsx
@@ -236,7 +236,7 @@ type BreadcrumbEllipsisOptions = {}
 export type BreadcrumbEllipsisProps = IconProps & BreadcrumbEllipsisOptions
 
 export const BreadcrumbEllipsis = forwardRef<BreadcrumbEllipsisProps, "span">(
-  ({ className, ...rest }, ref) => {
+  ({ children, className, ...rest }, ref) => {
     const styles = useBreadcrumb()
 
     const css: CSSUIObject = {
@@ -244,23 +244,25 @@ export const BreadcrumbEllipsis = forwardRef<BreadcrumbEllipsisProps, "span">(
     }
 
     return (
-      <Icon
-        ref={ref}
-        title="ellipsis"
-        className={cx("ui-breadcrumb__ellipsis", className)}
-        __css={css}
-        xmlns="http://www.w3.org/2000/svg"
-        viewBox="0 0 36 24"
-        stroke="currentColor"
-        strokeWidth="1.5"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-        {...rest}
-      >
-        <circle cx="10" cy="12" r="2" fill="currentColor" />
-        <circle cx="20" cy="12" r="2" fill="currentColor" />
-        <circle cx="30" cy="12" r="2" fill="currentColor" />
-      </Icon>
+      children ?? (
+        <Icon
+          ref={ref}
+          title="ellipsis"
+          className={cx("ui-breadcrumb__ellipsis", className)}
+          __css={css}
+          xmlns="http://www.w3.org/2000/svg"
+          viewBox="0 0 36 24"
+          stroke="currentColor"
+          strokeWidth="1.5"
+          strokeLinecap="round"
+          strokeLinejoin="round"
+          {...rest}
+        >
+          <circle cx="10" cy="12" r="2" fill="currentColor" />
+          <circle cx="20" cy="12" r="2" fill="currentColor" />
+          <circle cx="30" cy="12" r="2" fill="currentColor" />
+        </Icon>
+      )
     )
   },
 )