+++ packages/components/carousel/src/carousel-indicators.tsx
@@ -78,7 +78,7 @@ const CarouselIndicator: FC<CarouselIndicatorProps> = ({
     <ui.button
       type="button"
       tabIndex={-1}
-      className={cx("ui-carousel__indicators__indicator", className)}
+      className={cx("ui-carousel__indicator", className)}
       __css={css}
       {...rest}
     />