+++ packages/components/markdown/src/utils.ts
@@ -1,5 +1,5 @@
-import type { Break, Node, Paragraph, Text } from "mdast"
-import type { Literal, Parent } from "unist"
+import { isNull } from "@yamada-ui/utils"
+import type { Break, Node, Text } from "mdast"
 import { is } from "unist-util-is"
 import { getFragmentPattern } from "./patterns"
 
@@ -9,29 +9,11 @@ export interface ShouldRemoved extends Node {
 }
 
 const isObject = (target: unknown): target is { [key: string]: unknown } => {
-  return typeof target === "object" && target !== null
+  return typeof target === "object" && !isNull(target)
 }
 
 export const isNode = (node: unknown): node is Node => {
   return isObject(node) && "type" in node
-}
-
-export const isParent = (node: unknown): node is Parent => {
-  return isObject(node) && Array.isArray(node.children)
-}
-
-export const isLiteral = (node: unknown): node is Literal => {
-  return isObject(node) && "value" in node
-}
-
-export const isParagraph = (node: unknown): node is Paragraph => {
-  return isNode(node) && node.type === "paragraph"
-}
-
-export const isText = (node: unknown): node is Text => {
-  return (
-    isLiteral(node) && node.type === "text" && typeof node.value === "string"
-  )
 }
 
 export const isBreak = (node: unknown): node is Break => {
@@ -47,12 +29,8 @@ export const isContainer = (nodes: Node[]): boolean => {
     return false
   }
 
-  const firstNode = nodes.at(0)!
-  const lastNode = nodes.at(-1)!
-
-  if (!isText(firstNode) || !isText(lastNode)) {
-    return false
-  }
+  const firstNode = nodes.at(0)! as Text
+  const lastNode = nodes.at(-1)! as Text
 
   const hasBeginningPattern =
     getFragmentPattern("start", false).test(firstNode.value) &&